Move download CSV handler to parent Portfolio component
