+++ packages/client/src/ui/portfolio/CeloPortfolio.tsx
@@ -120,6 +120,7 @@ type ConnectProps = ReturnType<typeof mapStateToProps> & typeof dispatchProps;
 
 interface ComponentProps {
   fullSize: boolean;
+  downloadDataToFile: (CSV: string) => void;
 }
 
 interface IProps
